# api/openapi3.yaml
openapi: 3.0.0
info:
  version: 0.1.0
  title: Haiboo Api
  description: |
    Haiboo API POC
  contact:
    name: Imre Juhász
    url: 'https://github.com/mike-zenith'
    email: ~~@~~.com
servers:
  - url: 'http://localhost:9000'
    description: dev-localhost
  - url: 'http://127.0.0.1:9001'
    description: mock-localhost
paths:
  /productions:
    get:
      operationId: getProductions
      responses:
        '200':
          description: Successfuly retrieved productions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Productions'
              examples:
                example-all-productions:
                  value:
                    Items:
                      - Abstract: 'Három esztendővel azután, hogy Gotham maszkos igazságosztója eltűnt, Bruce Wayne unokatestvére, Kate Kane visszatér a városba, Hogy legyőzze démonait, magára ölti Batwoman álcáját, és Batman nyomdokaiba lépve szembeszáll a bűnözőkkel.'
                        AgeRating: 12
                        AvailabilityFromUtcIso: '2019-12-21T23:00:00Z'
                        BackgroundUrl: 'http://hboce-preprod-vod-hss.akamaized.net/d44b68b6-9fa1-f93d-9a84-84b3a941dc98_p106585_hbo/images/30445957.jpg'
                        Cast: 'Ruby Rose, Rachel Skarsten, Meagan Tandy, Camrus Johnson, Nicole Kang'
                        Category: SERIES
                        Director: Marcos Siega
                        EditedAbstract: 'Három esztendővel azután, hogy Gotham maszkos igazságosztója eltűnt, Bruce Wayne unokatestvére, Kate Kane visszatér a városba, Hogy legyőzze démonait, magára ölti Batwoman álcáját, és Batman nyomdokaiba lépve szembeszáll a bűnözőkkel.'
                        Genre: akció
                        Id: 604b537a-5ddb-2cd4-9c4e-8cccd248831d
                        Name: Batwoman
                        ProductionYear: 2019
                      - Abstract: A sportkörnyezetben játszódó afroamerikai tinidráma Spencer Paysinger amerikai futballjátékos igaz történetét meséli el. A Los Angeles egyik gettónegyedében élő Spencer a gazdag Beverly Hills középiskola csapatához igazol.
                        AgeRating: 12
                        AvailabilityFromUtcIso: '2018-10-11T07:00:00Z'
                        BackgroundUrl: 'https://cdn.hbogo.eu/images/D3D07ADF-7EFF-4146-AF3D-A36E113395F8/1920_1080.jpg'
                        Cast: 'Daniel Ezra, Greta Onieogou, Samantha Logan, Michael Evans Behling, Taye Diggs'
                        Category: SERIES
                        Director: Rob Hardy
                        EditedAbstract: A sportkörnyezetben játszódó afroamerikai tinidráma Spencer Paysinger amerikai futballjátékos igaz történetét meséli el. A Los Angeles egyik gettónegyedében élő Spencer a gazdag Beverly Hills középiskola csapatához igazol.
                        Genre: dráma
                        Id: e74eb909-2d8d-4495-b1de-b3da9e2d4b2e
                        Name: Spencer
                        ProductionYear: 2018
                      - Abstract: 'A sorozat az amerikai pornóipar felemelkedését meséli el az 1970-es, 80-as években. A pornó legalizálása után a New York-i Time Square lesz a műfaj egyik fellegvára, ahol a pénz mellett megjelenik a prostitúció, drog, AIDS és a maffia is.'
                        AgeRating: 18
                        AvailabilityFromUtcIso: '2017-09-23T22:00:00Z'
                        BackgroundUrl: 'http://level3cdn.hbogo.eu/d7e365d8-3e5d-456f-afa5-2c8acb614487_hbo/images/30112077.jpg'
                        Cast: 'James Franco, Maggie Gyllenhaal, Gary Carr, Margarita Levieva, Lawrence Gilliard Jr.'
                        Category: SERIES
                        Director: Michelle Maclaren
                        EditedAbstract: 'A sorozat az amerikai pornóipar felemelkedését meséli el az 1970-es, 80-as években. A pornó legalizálása után a New York-i Time Square lesz a műfaj egyik fellegvára, ahol a pénz mellett megjelenik a prostitúció, drog, AIDS és a maffia is.'
                        Genre: dráma
                        Id: d7e365d8-3e5d-456f-afa5-2c8acb614487
                        Name: Fülledt utcák
                        ProductionYear: 2017
                      - Abstract: 'A széria a világsikert arató Vámpírnaplók és A sötétség kora című vámpíros sorozatok folytatását meséli el, egy új természetfölötti képességekkel rendelkező generáció történetét, a középpontban Klaus Mikkelson 17 éves lányával, Hope-pal.'
                        AgeRating: 15
                        AvailabilityFromUtcIso: '2018-10-26T07:00:00Z'
                        BackgroundUrl: 'https://cdn.hbogo.eu/images/4701180F-B3F5-477C-9C78-97D9830A10FD/1080_608.jpg'
                        Cast: 'Danielle Rose Russell, Matthew Davis, Jenny Boyd, Kaylee Bryant, Quincy Fouse'
                        Category: SERIES
                        Director: Chris Grismer
                        EditedAbstract: 'A széria a világsikert arató Vámpírnaplók és A sötétség kora című vámpíros sorozatok folytatását meséli el, egy új természetfölötti képességekkel rendelkező generáció történetét, a középpontban Klaus Mikkelson 17 éves lányával, Hope-pal.'
                        Genre: dráma
                        Id: d8fad5bd-716d-9f55-0f0d-cd6f71864c5b
                        Name: Legacies - A sötétség öröksége
                        ProductionYear: 2018
                      - Abstract: 'Eszter a valóság elől a képregények világába menekül, csak az angol tanárát kedveli. Élete felfordul, amikor a tanár bejelenti, hogy külföldre megy dolgozni. Eszter különleges búcsú e-mailje után titkos távkapcsolat alakul ki közöttük.'
                        AgeRating: 12
                        AvailabilityFromUtcIso: '2019-06-05T22:00:00Z'
                        BackgroundUrl: 'http://hboce-preprod-vod-hss.akamaized.net/d88b3e44-df5b-3d05-0f26-a4be6ab9edb7_hbo/images/29058139.jpg'
                        Cast: 'Szilvia Herr, Kristóf Vajda, Csaba Polgár, Judit Rezes, Dávid Rácz'
                        Category: MOVIE
                        Director: Mihály Schwechtje
                        EditedAbstract: 'Eszter a valóság elől a képregények világába menekül, csak az angol tanárát kedveli. Élete felfordul, amikor a tanár bejelenti, hogy külföldre megy dolgozni. Eszter különleges búcsú e-mailje után titkos távkapcsolat alakul ki közöttük.'
                        Genre: dráma
                        Id: d88b3e44-df5b-3d05-0f26-a4be6ab9edb7
                        Name: 'Remélem legközelebb sikerül meghalnod :)'
                        ProductionYear: 2018
      description: Get all productions
      summary: ''
      tags:
        - Production
        - Collection
        - Read
  '/productions/{id}':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: |
          id of the production
    get:
      summary: ''
      tags:
        - Production
        - Read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionWithRequired'
            application/xml:
              schema:
                type: object
                properties: {}
        '404':
          description: Not Found
      operationId: getProduction
      description: Get a single production by id
    put:
      summary: ''
      operationId: createProduction
      responses:
        '200':
          description: 'OK, Existing modified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionWithRequired'
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionWithRequired'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description: Create or update (overwrite) production by id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionWithRequired'
        description: The production that has to be put under id
      tags:
        - Production
        - Write
    patch:
      summary: ''
      operationId: updateProduction
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionWithRequired'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
      description: Update production
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Production'
        description: Partial body of the production that has to be updated
      tags:
        - Production
        - Write
    delete:
      summary: ''
      operationId: deleteProduction
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
      description: Delete a production
      tags:
        - Production
        - Write
components:
  schemas:
    Productions:
      type: object
      description: Represents a collection of productions
      properties:
        Items:
          type: array
          items:
            $ref: '#/components/schemas/ProductionWithRequired'
    Production:
      type: object
      properties:
        Abstract:
          x-faker: random.words
          type: string
        AgeRating:
          type: integer
          minimum: 1
        AvailabilityFromUtcIso:
          x-faker: date.past
          type: string
          format: date-time
        BackgroundUrl:
          x-faker: image.city
          type: string
          format: uri
        Cast:
          x-faker: random.words
          type: string
        Category:
          type: string
          enum:
            - SERIES
            - MOVIE
        Director:
          x-faker: random.words
          type: string
        EditedAbstract:
          x-faker: random.words
          type: string
        Genre:
          type: string
        Id:
          x-faker: random.uid
          type: string
          format: uuid
        Name:
          x-faker: name.findName
          type: string
        ProductionYear:
          x-faker:
          type: integer
          format: date
          minimum: 1888
          maximum: 2021
      description: Represents a single production
    ProductionRequiredProp:
      type: object
      required:
        - Abstract
        - AgeRating
        - AvailabilityFromUtcIso
        - BackgroundUrl
        - Cast
        - Category
        - Director
        - EditedAbstract
        - Genre
        - Id
        - Name
        - ProductionYear
    Error:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
              keyword:
                type: string
              schemaPath:
                type: string
            required:
              - message
              - keyword
              - schemaPath
      required:
        - errors
      x-examples:
        example-error: |
          {
            "errors": [{
              keyword: 'required',
              schemaPath: '#/allOf/0/required',
              message: "should have required property 'Genre'"
            }]
          }
    ProductionWithRequired:
      title: PrudctionWithRequired
      allOf:
        - $ref: '#/components/schemas/ProductionRequiredProp'
        - $ref: '#/components/schemas/Production'
tags:
  - name: Production
  - name: Collection
  - name: Read
  - name: Write